+++ src/main/java/br/com/gasoutapp/dto/SensorDetailsDTO.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class SensorDetailsDTO {
 	private String name;
-  private boolean alarmOn;
+	private boolean alarmOn;
 	private boolean notificationOn;
-	private boolean sprinklersOn;	
-	private Integer sensorValue;
+	private boolean sprinklersOn;
+	private Long sensorValue;
 }
